@@ -39,17 +39,21 @@
   },
   "dependencies": {
     "chart.js": "^2.1.1",
+    "draft-js": "^0.7.0",
+    "firebase": "^3.0.1",
     "google-maps": "^3.2.1",
     "history": "^1.17.0",
     "material-ui": "^0.15.0",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
     "react-helmet": "^3.1.0",
     "react-hot-loader": "^1.3.0",
     "react-redux": "^4.4.5",
     "react-router": "^2.4.0",
     "react-router-scroll": "^0.2.0",
+    "react-share": "^1.8.0",
     "react-tap-event-plugin": "^1.0.0",
+    "reactfire": "^1.0.0",
     "redux": "^3.5.2",
     "socket.io-client": "^1.4.6"
   }
