@@ -1,16 +1,9 @@
-import React from 'react';
+import Button from 'material-ui/FlatButton';
+import doctors from '../data/doctors';
+import DoctorsIcon from 'material-ui/svg-icons/navigation/arrow-back';
 import Helmet from 'react-helmet';
 import { Link } from 'react-router';
-import Button from 'material-ui/FlatButton';
-import DoctorsIcon from 'material-ui/svg-icons/navigation/arrow-back';
-import * as Colors from 'material-ui/styles/colors';
-import doctors from '../data/doctors';
-
-const styles = {
-  header : {
-    textTransform:'capitalize',
-  },
-};
+import React from 'react';
 
 function getDoctor(id) {
   for (var i = doctors.length - 1; i >= 0; i--) {
@@ -20,29 +13,33 @@ function getDoctor(id) {
   }
 }
 
-export default React.createClass({
-  contextTypes: {
-    store:  React.PropTypes.object,
-    router: React.PropTypes.object,
-  },
+class Doctor extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleTouchTap = this.handleTouchTap.bind(this);
+  }
+
   handleTouchTap() {
     const route = "/doctors", index = 3;
     // this.context.router.push(route);
     this.context.store.dispatch({type:"UPDATE_ROUTE", route});
     this.context.store.dispatch({type:"UPDATE_INDEX", index});
-  },
+  }
+
   render() {
+    const { primary1Color, textColor } = this.context.muiTheme.palette;
     const { id } = this.props.params;
     const doctor = getDoctor(id);
-    const description = doctor.description.map((p, i) => (
-      <div key={i} className="doctor-text"
-      	dangerouslySetInnerHTML={{__html:p}}/>
+    const description = doctor.description.map((section, i) => (
+      <p key={i} dangerouslySetInnerHTML={{__html: section}}/>
     ));
+
     return (
       <article>
-        <Helmet 
-          title={"Doctor " + doctor.firstname + " " + doctor.lastname} 
+        <Helmet
+          title = {"Doctor " + doctor.firstname + " " + doctor.lastname}
         />
+
         <div style={{textAlign:'left'}}>
           <Button className="button-top-left"
             label="All Doctors" 
@@ -50,19 +47,29 @@ export default React.createClass({
             primary={true}
             icon={<DoctorsIcon />}
             linkButton={true}
-            rippleColor={Colors.pink500}
             containerElement={<Link to="/doctors" />}
             onTouchTap={this.handleTouchTap}
           />
         </div>
 
-        <h4 style={styles.header}>
+        <h3 style={{textTransform:'capitalize'}}>
           Doctor {doctor.firstname} {doctor.lastname}
-        </h4>
+        </h3>
 
-        <img width="100%" src={require("../images/doctors/" + doctor.img.big)}/>
-        {description}
+        <img width="100%" 
+          src={require("../images/doctors/" + doctor.img.big)}
+        />
+
+        <div style={{margin:50, color:textColor}}>{description}</div>
       </article>
     )
   }
-});
+}
+
+Doctor.contextTypes = {
+  store:  React.PropTypes.object,
+  router: React.PropTypes.object,
+  muiTheme: React.PropTypes.object,
+};
+
+export default Doctor;
