@@ -1,44 +1,61 @@
-import React       from 'react';
-import Helmet      from 'react-helmet';
-import Card        from 'material-ui/Card';
-import CardMedia   from 'material-ui/Card/CardMedia';
-import CardTitle   from 'material-ui/Card/CardTitle';
-import CardText    from 'material-ui/Card/CardText';
-import List        from 'material-ui/List';
-import ListItem    from 'material-ui/List/ListItem';
-import * as Colors from 'material-ui/styles/colors';
-import Divider     from '../components/ui/ListItemDivider';
-import services    from '../data/services';
+import Card from 'material-ui/Card';
+import CardMedia from 'material-ui/Card/CardMedia';
+import CardTitle from 'material-ui/Card/CardTitle';
+import CardText from 'material-ui/Card/CardText';
+import Divider from '../components/ui/ListItemDivider';
+import Helmet from 'react-helmet';
+import List from 'material-ui/List';
+import ListItem from 'material-ui/List/ListItem';
+import React from 'react';
+import services from '../data/services';
 
-export default () => (
-  <section>
-    <Helmet title="Services - Metromed UC"/>
+class Services extends React.Component {
+  constructor(props) {
+    super(props);
+  }
 
-    <h3>Services</h3>
+  render() {
+    const { primary2Color, accent3Color, textColor } = this.context.muiTheme.palette;
+    return (
+      <section>
+        <Helmet title="Services - Metromed UC"/>
 
-    <div className="flex-container">
-      {services.map((service, i) => (
-        <div key={i} className="flex-item">
-          <Card style={{textAlign:'center'}}>
-            <CardTitle style={{fontSize:34, color:Colors.cyan700}}>
-              <span>{service.title}</span>
-            </CardTitle>
-            <CardMedia style={{margin:0, padding:0, fontSize:25}}>
-              <span style={{fontSize:77, color:Colors.grey700}}>${service.price}</span>
-            </CardMedia>
-            <CardText>
-              <List>
-                {service.list.map((li, i, arr) => (
-                  <ListItem key={i}>
-                    <span style={{fontSize:18, color:Colors.blueGrey700}}>{li}</span>
-                    <Divider style={{marginLeft:20, marginRight:20, transform:'translateY(16px)'}} index={i} length={arr.length}/>
-                  </ListItem>
-                ))}
-              </List>
-            </CardText>
-          </Card>
+        <h3>Services</h3>
+
+        <div className="flex-container">
+          {services.map((service, i) => (
+            <div key={i} className="flex-item">
+              <Card style={{textAlign:'center'}}>
+                <CardTitle style={{fontSize:34, color:primary2Color}}>
+                  <span>{service.title}</span>
+                </CardTitle>
+
+                <CardMedia style={{margin:0, padding:0, fontSize:25}}>
+                  <span style={{fontSize:77, color:primary2Color}}>${service.price}</span>
+                </CardMedia>
+
+                <CardText>
+                  <List>
+                    {service.list.map((li, i, arr) => (
+                      <ListItem key={i} disabled>
+                        <span style={{fontSize:18, color:textColor}}>{li}</span>
+
+                        <Divider style={{marginLeft:20, marginRight:20, transform:'translateY(16px)'}} index={i} length={arr.length}/>
+                      </ListItem>
+                    ))}
+                  </List>
+                </CardText>
+              </Card>
+            </div>
+          ))}
         </div>
-      ))}
-    </div>
-  </section>
-);
+      </section>
+    );
+  }
+}
+
+Services.contextTypes = {
+  muiTheme: React.PropTypes.object.isRequired,
+};
+
+export default Services;
